+++ API/project/src/main/java/com/grademanager/demo/service/CourseService.java
@@ -21,7 +21,7 @@ public class CourseService {
         return courseRepository.save(course);
     }
 
-    public Course getCourse(Integer id){
+    public Course getCourse(Long id){
       Optional <Course> optionalCourse = courseRepository.findById(id);
         if(!optionalCourse.isPresent()){
             String err = String.format("The course having ID %s was not found", id);
